Added a user style filter based on Hunter.io.
This commit includes a significant number of comments being added, and
also fixes several small bugs that were present in the user scan.
Additional output was added for clarification and so the user can
roughly know how long the user scan will take once the HaveIBeenPwned
full scan starts.

Additionally we're now reporting which breaches a user is in, and the
compromised emails + breaches are using JSON (the format might need some
adjustment).