Switched domain scan to JSON, fixed wpscan
This is the first set of changes to use JSON as our output format. These
changes are currently restricted to the core domain scan and the related
filters. Doing the changes progressively lets us verify that everything
still works in smaller chunks. Seeing this output is suggesting that
we'll be able to combine into a single JSON document at some point.

wpscan had some pathing issues, fixed those up.