Added an extensible filter over the name data.
See the file user-scan/name-bad-word-list for the current list of items
(one-per-line) stripped out from names. Removals are literal and case
sensitive so we should account for that. Also be careful since the words
are literal we might unintentionally strip out a real part of a name

One thing we could add in the future is a way to remove isolated letters
and such.